Component Refactor
Brings navigation and theme components into their own folder within $lib.
