+++ src/lib/components/ui/navigation/NavbarLinkMobile.svelte
