+++ src/lib/components/ui/navigation/NavigationClose.svelte
